+++ lib/mcscf/fci_string.c
@@ -14,16 +14,16 @@
  * Hamming weight popcount
  */
 
-int FCIpopcount_1(unsigned long x)
+int FCIpopcount_1(size_t x)
 {
-        const unsigned long m1  = 0x5555555555555555; //binary: 0101...
-        const unsigned long m2  = 0x3333333333333333; //binary: 00110011..
-        const unsigned long m4  = 0x0f0f0f0f0f0f0f0f; //binary:  4 zeros,  4 ones ...
-        const unsigned long m8  = 0x00ff00ff00ff00ff; //binary:  8 zeros,  8 ones ...
-        const unsigned long m16 = 0x0000ffff0000ffff; //binary: 16 zeros, 16 ones ...
-        const unsigned long m32 = 0x00000000ffffffff; //binary: 32 zeros, 32 ones
-//        const unsigned long hff = 0xffffffffffffffff; //binary: all ones
-//        const unsigned long h01 = 0x0101010101010101; //the sum of 256 to the power of 0,1,2,3...
+        const size_t m1  = 0x5555555555555555; //binary: 0101...
+        const size_t m2  = 0x3333333333333333; //binary: 00110011..
+        const size_t m4  = 0x0f0f0f0f0f0f0f0f; //binary:  4 zeros,  4 ones ...
+        const size_t m8  = 0x00ff00ff00ff00ff; //binary:  8 zeros,  8 ones ...
+        const size_t m16 = 0x0000ffff0000ffff; //binary: 16 zeros, 16 ones ...
+        const size_t m32 = 0x00000000ffffffff; //binary: 32 zeros, 32 ones
+//        const size_t hff = 0xffffffffffffffff; //binary: all ones
+//        const size_t h01 = 0x0101010101010101; //the sum of 256 to the power of 0,1,2,3...
         x = (x & m1 ) + ((x >>  1) & m1 ); //put count of each  2 bits into those  2 bits 
         x = (x & m2 ) + ((x >>  2) & m2 ); //put count of each  4 bits into those  4 bits 
         x = (x & m4 ) + ((x >>  4) & m4 ); //put count of each  8 bits into those  8 bits 
@@ -33,7 +33,7 @@ int FCIpopcount_1(unsigned long x)
         return x;
 }
 
-int FCIpopcount_4(unsigned long x)
+int FCIpopcount_4(size_t x)
 {
         int count;
         for (count = 0; x; count++) {
@@ -43,9 +43,9 @@ int FCIpopcount_4(unsigned long x)
 }
 
 
-int FCIparity(unsigned long string0, unsigned long string1)
+int FCIparity(size_t string0, size_t string1)
 {
-        unsigned long ss;
+        size_t ss;
         if (string1 > string0) {
                 ss = string1 - string0;
                 // string1&ss gives the number of 1s between two strings
@@ -69,8 +69,8 @@ int FCIparity(unsigned long string0, unsigned long string1)
 static int binomial(int n, int m)
 {
         int i;
-        unsigned long num = 1;
-        unsigned long div = 1;
+        size_t num = 1;
+        size_t div = 1;
         double dnum = 1;
         double ddiv = 1;
         if (n < 28) {
@@ -102,7 +102,7 @@ static int binomial(int n, int m)
         }
 }
 
-int FCIstr2addr(int norb, int nelec, unsigned long string)
+int FCIstr2addr(int norb, int nelec, size_t string)
 {
         int addr = 0;
         int nelec_left = nelec;
@@ -120,14 +120,14 @@ int FCIstr2addr(int norb, int nelec, unsigned long string)
 
 
 void FCIlinkstr_index(int *link_index, int norb, int na, int nocc,
-                      unsigned long *strs, int store_trilidx)
+                      size_t *strs, int store_trilidx)
 {
         int occ[norb];
         int vir[norb];
         int nvir = norb - nocc;
         int nlink = nocc * nvir + nocc;
         int str_id, io, iv, i, a, k, ia;
-        unsigned long str0, str1;
+        size_t str0, str1;
         int *tab;
 
         for (str_id = 0; str_id < na; str_id++) {
